renaming WebApi.Tests
