+++ NETCORE/ApplicationInsights.AspNetCore.sln
@@ -28,8 +28,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.ApplicationInsigh
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.ApplicationInsights.AspNetCore.Tests", "test\Microsoft.ApplicationInsights.AspNetCore.Tests\Microsoft.ApplicationInsights.AspNetCore.Tests.csproj", "{2766D8AF-C20B-4F3A-8260-6C2D39B7A8A0}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApi20.FunctionalTests20", "test\WebApi20.FunctionalTests\WebApi20.FunctionalTests20.csproj", "{C3B3F515-0305-4809-A9A8-37FD80428F74}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ApplicationInsightsTypes", "test\ApplicationInsightsTypes\ApplicationInsightsTypes.csproj", "{9DA7024F-216F-4FA5-9B6D-CE4216C2DD72}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IntegrationTests.WebApp", "test\IntegrationTests.WebApp\IntegrationTests.WebApp.csproj", "{8E71FECF-E090-409E-8551-C597F9DFB91C}"
@@ -64,6 +62,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FunctionalTests.Utils", "te
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FunctionalTests.MVC.Tests", "test\FunctionalTests.MVC.Tests\FunctionalTests.MVC.Tests.csproj", "{9693CC30-A3E6-48A7-A68E-6E0DD098D779}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FunctionalTests.WebApi.Tests", "test\FunctionalTests.WebApi.Tests\FunctionalTests.WebApi.Tests.csproj", "{E78404EA-CBAA-4366-B659-C3A46BC9DF15}"
+EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
 		..\BASE\src\Common\Common\Common.projitems*{0c91d961-b21e-4760-a192-14193f10b831}*SharedItemsImports = 5
@@ -94,10 +94,6 @@ Global
 		{2766D8AF-C20B-4F3A-8260-6C2D39B7A8A0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2766D8AF-C20B-4F3A-8260-6C2D39B7A8A0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2766D8AF-C20B-4F3A-8260-6C2D39B7A8A0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C3B3F515-0305-4809-A9A8-37FD80428F74}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C3B3F515-0305-4809-A9A8-37FD80428F74}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C3B3F515-0305-4809-A9A8-37FD80428F74}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C3B3F515-0305-4809-A9A8-37FD80428F74}.Release|Any CPU.Build.0 = Release|Any CPU
 		{9DA7024F-216F-4FA5-9B6D-CE4216C2DD72}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9DA7024F-216F-4FA5-9B6D-CE4216C2DD72}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9DA7024F-216F-4FA5-9B6D-CE4216C2DD72}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -158,6 +154,10 @@ Global
 		{9693CC30-A3E6-48A7-A68E-6E0DD098D779}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9693CC30-A3E6-48A7-A68E-6E0DD098D779}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9693CC30-A3E6-48A7-A68E-6E0DD098D779}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E78404EA-CBAA-4366-B659-C3A46BC9DF15}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E78404EA-CBAA-4366-B659-C3A46BC9DF15}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E78404EA-CBAA-4366-B659-C3A46BC9DF15}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E78404EA-CBAA-4366-B659-C3A46BC9DF15}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -166,7 +166,6 @@ Global
 		{16B44D67-6214-4DDE-B419-93EF073E2E69} = {8B5230E5-8138-44D6-839F-DF9248F195EE}
 		{95EC3635-22E4-4C3A-A066-F5823A0648DA} = {2E6DDE9E-8C75-4F9C-8906-08EBDD6E73EF}
 		{2766D8AF-C20B-4F3A-8260-6C2D39B7A8A0} = {8B5230E5-8138-44D6-839F-DF9248F195EE}
-		{C3B3F515-0305-4809-A9A8-37FD80428F74} = {8B5230E5-8138-44D6-839F-DF9248F195EE}
 		{9DA7024F-216F-4FA5-9B6D-CE4216C2DD72} = {8B5230E5-8138-44D6-839F-DF9248F195EE}
 		{8E71FECF-E090-409E-8551-C597F9DFB91C} = {8B5230E5-8138-44D6-839F-DF9248F195EE}
 		{FE9DB9A7-D9AE-4188-945C-393D70022D9A} = {8B5230E5-8138-44D6-839F-DF9248F195EE}
@@ -183,6 +182,7 @@ Global
 		{C7A8D822-CA6C-497D-AA26-C36F25BA61FD} = {8B5230E5-8138-44D6-839F-DF9248F195EE}
 		{63F64005-CDC7-4062-8CBC-A40B049E342A} = {8B5230E5-8138-44D6-839F-DF9248F195EE}
 		{9693CC30-A3E6-48A7-A68E-6E0DD098D779} = {8B5230E5-8138-44D6-839F-DF9248F195EE}
+		{E78404EA-CBAA-4366-B659-C3A46BC9DF15} = {8B5230E5-8138-44D6-839F-DF9248F195EE}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {047855A4-470F-43B1-8B74-69651DD6B8A6}
