+++ NETCORE/test/FunctionalTests.WebApi.Tests/Controllers/DependencyController.cs
@@ -5,7 +5,7 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 
-namespace WebApi20.FunctionalTests.Controllers
+namespace FunctionalTests.WebApi.Tests.Controllers
 {
     [Route("api/[controller]")]
     public class DependencyController : Controller
