+++ NETCORE/test/FunctionalTests.WebApi.Tests/FunctionalTest/RequestCollectionTests.cs
@@ -1,4 +1,4 @@
-﻿namespace WebApi20.FuncTests
+﻿namespace FunctionalTests.WebApi.Tests.FunctionalTest
 {
     using System;
     using System.Collections.Generic;
