+++ NETCORE/test/FunctionalTests.WebApi.Tests/Properties/launchSettings.json
