+++ NETCORE/test/FunctionalTests.WebApi.Tests/appsettings.json
